requirements.txt
fastapi[standard]
uvicorn
python-dotenv
google-generativeai